minor edit and add README
